@@ -3,6 +3,7 @@
   "description": "prisma api with authentication",
   "version": "1.0.0",
   "author": "Piers Pollard",
+  "url": "www.ourtrees.eco",
   "private": true,
   "dependencies": {
     "@apollo/server": "^4.9.3",
@@ -40,29 +41,33 @@
     "request": "^2.88.2",
     "ts-essentials": "^7.0.1",
     "type-graphql": "^2.0.0-beta.1",
-    "typescript": "^5.2.2",
+    "typescript": "^5.3.2",
     "uuid": "^3.3.3",
     "uuid4": "^2.0.3"
   },
   "devDependencies": {
     "@types/express": "^4.17.18",
     "@types/fs-extra": "^9.0.1",
-    "@types/jest": "^26.0.19",
+    "@types/jest": "^29.5.10",
     "@types/minio": "^7.0.5",
     "@types/mustache": "^4.0.1",
     "@types/sendgrid": "^4.3.0",
     "@types/sinon": "^9.0.10",
-    "@typescript-eslint/eslint-plugin": "^2.8.0",
-    "@typescript-eslint/parser": "^2.8.0",
+    "@typescript-eslint/eslint-plugin": "^6.13.1",
+    "@typescript-eslint/parser": "^6.13.1",
     "browser-sync": "^2.26.7",
-    "eslint": "^6.6.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jest": "^24.1.3",
+    "eslint": "^8.54.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jest": "^27.6.0",
+    "eslint-plugin-no-only-tests": "^3.1.0",
+    "eslint-plugin-testing-library": "^6.2.0",
     "favicons": "^6.2.0",
     "fetch-cookie": "^0.11.0",
+    "form-data": "^4.0.0",
     "graphql-request": "^6.1.0",
     "jest": "^29.7.0",
+    "node-fetch": "^3.3.2",
     "nodemon": "^2.0.0",
     "prettier": "3.1.0",
     "random-utility": "^1.0.1",
