@@ -2,13 +2,12 @@ import { gql } from 'graphql-request';
 import '../../global-variables';
 import path from 'path';
 import { createReadStream } from 'fs-extra';
+import FormData from 'form-data';
+import fetch from 'node-fetch';
 import TestUtils from '../utils';
 import { CreateTreeEntryInput } from '../../src/resolvers/mutation/createTreeEntry';
 import { FileHandler } from '../../src/modules/FileHandler';
 import { TreeEntryProfile } from '../../src/types/TreeEntryProfile';
-
-import FormData from 'form-data';
-import fetch from 'node-fetch';
 
 const query = gql`
   mutation createTreeEntry($data: CreateTreeEntryInput!) {
@@ -128,7 +127,7 @@ test('should fail if user does not own tree', async () => {
     );
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Unauthorized');
   }
 });
@@ -152,7 +151,7 @@ test('should fail if tree does not exist', async () => {
     );
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Tree does not exist');
   }
 });
