@@ -81,7 +81,7 @@ test('should fail to delete notification that requestor is not a receiver of', a
     );
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Unauthorised');
   }
 });
@@ -97,7 +97,7 @@ test('should fail to delete notification that does not exist', async () => {
     );
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Unauthorised');
   }
 });
