@@ -30,7 +30,7 @@ test('should fail - User Does Not Exist', async () => {
     });
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('User Does Not Exist');
   }
 });
