@@ -61,7 +61,7 @@ test('should fail - Token Invalid', async () => {
     });
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Token Invalid');
   }
 });
