@@ -34,7 +34,7 @@ test('should succeed', async () => {
   expect(data?.verifyUser.user.verified).toBeTruthy();
 });
 
-test("should fail - user doesn't exist", async () => {
+test('should fail - user doesn\'t exist', async () => {
   const token = generateToken({
     type: TokenType.VERIFICATION,
     data: {
@@ -48,7 +48,7 @@ test("should fail - user doesn't exist", async () => {
     });
     throw new Error();
   } catch (error) {
-    // eslint-disable-next-line jest/no-conditional-expect, jest/no-try-expect
+    // eslint-disable-next-line jest/no-conditional-expect
     expect(error.response.errors[0].message).toEqual('Unauthorised');
   }
 });
