@@ -4,7 +4,7 @@ import { Prisma, PrismaClient } from '@prisma/client';
 /**
  * Remap prisma instances and omit read and write actions appropriately
  */
-export interface Prisma {
+export interface PrismaInterface {
   write: DeepOmit<
     PrismaClient,
     {
@@ -34,7 +34,7 @@ export interface Prisma {
   >;
 }
 
-export default (): Prisma => ({
-  write: new PrismaClient({ datasources: { db: { url: process.env.DATABASE_URL_WRITE } } }) as any,
-  read: new PrismaClient({ datasources: { db: { url: process.env.DATABASE_URL_READ } } }) as any,
+export default (): PrismaInterface => ({
+  write: new PrismaClient({ datasources: { db: { url: process.env.DATABASE_URL_WRITE } } }),
+  read: new PrismaClient({ datasources: { db: { url: process.env.DATABASE_URL_READ } } }),
 });
